+++ lib/main的副本.dart
@@ -1,7 +1,86 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'dart:convert';
 
-void main() => runApp(MyApp());
+void main() {
+  runApp(new MyApp());
+}
+
+/**
+ * 垂直listView
+ */
+// class MyApp extends StatelessWidget {
+//   @override
+//   Widget build(BuildContext context) {
+//     final title = 'list';
+//     return new MaterialApp(
+//       title: title,
+//       home: new Scaffold(
+//         appBar: new AppBar(
+//           title: new Text(title),
+//         ),
+//         body: Column(
+//           mainAxisAlignment: MainAxisAlignment.end,
+//           children: <Widget>[
+//             Container(
+//               height: 100,
+//               color: Colors.red,
+//               child: ListView(
+//                 //控制方向 默认是垂直的
+//                 scrollDirection: Axis.horizontal,
+//                 padding: const EdgeInsets.all(16.0),
+//                 children: <Widget>[
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('phone', Icons.phone),
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('phone', Icons.phone),
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('phone', Icons.phone),
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('Maps', Icons.map),
+//                   _getContainer('phone', Icons.phone),
+//                   _getContainer('Maps', Icons.map),
+//                 ],
+//               ),
+//             ),
+//             Text(
+//               'You have pushed the button this many times:',
+//             ),
+//           ],
+//         ),
+//       ),
+//     );
+//   }
+
+/**
+   * 抽取item项
+   */
+//   Widget _getContainer(String test, IconData icon) {
+//     return new Container(
+//       height: 60.0,
+//       color: Colors.red,
+//       child: Column(
+//         mainAxisAlignment: MainAxisAlignment.center,
+//         children: <Widget>[
+//           Image(image: new AssetImage('images/ic_assignment_ind_36pt.png')),
+//           Text(
+//             'abc',
+//           ),
+//         ],
+//       ),
+//     );
+//   }
+// }
+
+class BasicMenuItem {
+  String name;
+  Image image;
+  String callbackname;
+  BasicMenuItem(this.name, this.image, this.callbackname);
+}
 
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
@@ -48,28 +127,81 @@ class _MyHomePageState extends State<MyHomePage> {
   static const EventChannel eventChannel =
       const EventChannel('com.pages.your/native_post');
 
+  var list = <Map>[];
+  var functionItems = <BasicMenuItem>[];
+
   // 渲染前的操作，类似viewDidLoad
   @override
   void initState() {
     super.initState();
 
     // 监听事件，同时发送参数12345
-    // eventChannel
-    //     .receiveBroadcastStream(12345)
-    //     .listen(_onEvent, onError: _onError);
+    eventChannel
+        .receiveBroadcastStream('init')
+        .listen(_onEvent, onError: _onError);
+
+    var nameList = [
+      '清屏',
+      '高清',
+      '举报',
+      '管理',
+      '互动',
+      '直播间管理',
+      '音频模式',
+      '子频道',
+      '私聊',
+      '最小化'
+    ];
+
+    var imageList = [
+      'images/mlive_setting_item_clearWindow.png',
+      'images/mlive_setting_item_hd.png',
+      'images/mlive_setting_item_report.png',
+      'images/mlive_setting_item_manage.png',
+      'images/mlive_setting_item_plugin.png',
+      'images/mlive_setting_item_manage.png',
+      'images/mlive_setting_item_audio.png',
+      'images/mlive_setting_item_subchannel.png',
+      'images/mlive_setting_item_whisper.png',
+      'images/mlive_setting_item_minimize.png'
+    ];
+
+    var funList = [
+      'clearWindow',
+      'hd',
+      'report',
+      'manage',
+      'plugin',
+      'manage',
+      'audio',
+      'subchannel',
+      'hisper',
+      'minimize'
+    ];
+
+    for (var i = 0; i < 10; i++) {
+      var image = Image(image: new AssetImage(imageList[i]));
+      var item = BasicMenuItem(nameList[i], image, funList[i]);
+      functionItems.add(item);
+    }
   }
 
   // 回调事件
   void _onEvent(Object event) {
-    setState(() {
-      _counter++;
-    });
+    Map<String, dynamic> user = jsonDecode(event);
+    // print("flutter方法调用" + user["first"]);
+    // print("flutter方法调用" + user["second"]);
+    list.add(user);
+    //print("flutter方法调用" + list.toString());
+    // setState(() {
+    //   _counter = _counter ++;
+    // });
   }
 
   // 错误返回
   void _onError(Object error) {}
 
-  int _counter = 0;
+  int _counter = 20;
 
   void _incrementCounter() {
     setState(() {
@@ -85,77 +217,81 @@ class _MyHomePageState extends State<MyHomePage> {
   /**
    * 抽取item项
    */
-  Widget _getContainer(String test, IconData icon) {
+  Widget _getContainer(BasicMenuItem item) {
     return new Container(
-      width: 160.0,
-      //      ListTile
-      child: new ListTile(
-        //       显示在title之前
-        leading: new Icon(icon),
-        title: new Text(test),
+      color: Colors.blue,
+      height: 60,
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+          item.image,
+          Text(
+            item.name,
+          ),
+        ],
       ),
     );
   }
 
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
     return Scaffold(
-      appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
+      // appBar: AppBar(
+      //   // Here we take the value from the MyHomePage object that was created by
+      //   // the App.build method, and use it to set our appbar title.
+      //   title: Text(widget.title),
+      // ),
+      backgroundColor: Colors.transparent,
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: <Widget>[
+          // Text(
+          //   'You have pushed the button this many times:',
+          // ),
+          Text(
+            '$_counter',
+            style: Theme.of(context).textTheme.display1,
+          ),
+          // Image(image: new AssetImage('images/ic_assignment_ind_36pt.png')),
+          // Image(image: new AssetImage('images/ic_assignment_ind_36pt.png')),
+          Container(
+              // child: ListView(
+              //   //控制方向 默认是垂直的
+              //   scrollDirection: Axis.horizontal,
+              //   itemExtent: 60,
+              //   padding: const EdgeInsets.all(16.0),
+              //   children: <Widget>[
+              //     // for (BasicMenuItem item in functionItems) {
+              //     //   _getContainer(item);
+              //     // }
+
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('phone', Icons.phone),
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('phone', Icons.phone),
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('phone', Icons.phone),
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('Maps', Icons.map),
+              //     // _getContainer('phone', Icons.phone),
+              //     // _getContainer('Maps', Icons.map),
+              //   ],
+              // ),
+              child: new ListView.builder(
+                  scrollDirection: Axis.horizontal,
+                  itemExtent: 60,
+                  itemCount: functionItems.length,
+                  padding: const EdgeInsets.all(16.0),
+                  itemBuilder: (context, i) {
+                    return _getContainer(functionItems[i]);
+                  }),
+              height: 400,
+              color: Colors.red),
+        ],
       ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.display1,
-            ),
-            Image(image: new AssetImage('images/ic_assignment_ind_36pt.png')),
-            Image(image: new AssetImage('images/ic_assignment_ind_36pt.png')),
-            // ListView(
-            //   scrollDirection: Axis.horizontal,
-            //   children: <Widget>[
-            //     _getContainer('Maps', Icons.map),
-            //     _getContainer('phone', Icons.phone),
-            //     _getContainer('Maps', Icons.map),
-            //   ],
-            // ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }
