+++ pubspec.yaml
@@ -27,6 +27,12 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
+  battery: ^0.3.0+1
+  flutter_boost: ^0.0.400
+  url_launcher: ^5.0.2
+  package_info: ^0.4.0+2
+  flutter_plugin_crashsdk:
+    git: https://git.yy.com/lupeihong/flutter-plugin-crashsdk.git
 
 dev_dependencies:
   flutter_test:
